+++ checksum-Elixir.Tursox.Native.exs
@@ -1,9 +1,0 @@
-%{
-  "libtursox_nif-v0.1.0-nif-2.16-aarch64-apple-darwin.so.tar.gz" => "sha256:0b5ea7d001ce0b4aeac6ab1fae1a6de29361996f6789b5b38b1bccb825ba7481",
-  "libtursox_nif-v0.1.0-nif-2.16-aarch64-unknown-linux-gnu.so.tar.gz" => "sha256:fe19f557feaa06ac32a586b7476dafc4518a712cee6c7061eb68a51665f80923",
-  "libtursox_nif-v0.1.0-nif-2.16-aarch64-unknown-linux-musl.so.tar.gz" => "sha256:e04abbf4e34762ae91adccdb1940790e9d8b781b277eee3503df414f5bd5590c",
-  "libtursox_nif-v0.1.0-nif-2.16-x86_64-apple-darwin.so.tar.gz" => "sha256:c197889a56d8e252fc727c6b2d09c5f93093842c0efdabd8496ff69c77a8b6a7",
-  "libtursox_nif-v0.1.0-nif-2.16-x86_64-unknown-linux-gnu.so.tar.gz" => "sha256:d436716de6beed43955f42ac9f10eeb14d1657f14f640c3d22df2e6cfb42982d",
-  "libtursox_nif-v0.1.0-nif-2.16-x86_64-unknown-linux-musl.so.tar.gz" => "sha256:dec3ab90da40923d7e7a4415915788675253300275431e0053760bb6786d833d",
-  "tursox_nif-v0.1.0-nif-2.16-x86_64-pc-windows-msvc.dll.tar.gz" => "sha256:7ca50b1344d40058ef8fe688b2acadb0f448b1c7a0f7d0fec674276a35c6ee57",
-}